+++ CHANGELOG.md
@@ -5,6 +5,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Public commands
+  - Assert-TlsProtocol
+  - Disable-TlsProtocol
+  - Enable-TlsProtocol
+  - Get-TlsProtocol
+  - Test-TlsNegotiation
+  - Test-TlsProtocol
+
 ### Changed
 
 - SChannelDsc